net/mlx5e: IPSec, Add support for ESP trailer removal by hardware
Current hardware decrypts and authenticates incoming ESP packets.
Subsequently, the software extracts the nexthdr field, truncates the
trailer and adjusts csum accordingly.

With this patch and a capable device, the trailer is being removed by the
hardware and the nexthdr field is conveyed via PET. This way we avoid both
the need to access the trailer (cache miss) and to compute its relative
checksum, which significantly improve the performance.

Signed-off-by: Yossi Kuperman <yossiku@mellanox.com>
Change-Id: I4ca12b41cf8317502c0efc46cada569bc245362a
Issue: None