net/mlx5: Add empty egress namespace to flow steering core
Currently, we don't support egress flow steering namespace in mlx5
flow steering core implementation. However, when we want to encrypt
a packet, we model it as a flow steering rule in the egress path.
To overcome this, we add an empty egress namespace to flow steering.
This namespace is initialized only when ipsec support exists.
In the future, this will grow to a full blown full steering
implementation, resembling the ingress path.

issue: 1087238
Change-Id: Id74a4bc36ebf5015d1392a4138c9f52ebe4072b4
Signed-off-by: Matan Barak <matanb@mellanox.com>