net/mlx5: Add notifier chain for flow steering
Sometimes, entities (like FPGA special transformations) need to get
notifications when a handle (rule with possibly a few destinations) is
added or deleted. These entities may alter the user's input or do some
other action as a result of adding such a rule.

Adding a per namespace notifier chain system. Registered entities are
notifier three times: before a rule is added, after the rule has been
added (or if adding this rule failed) and before a rule is deleted.
In order to tie theses events each handle is allocated with unique
per-FT number. This number is used to tie all these events.
Compiling this notifiers is optional, as it may reduce rules update
rate.

Adding API to register and unregister to namespace notifiers and
defining the required structs for notifications.

issue: 1087238
Change-Id: I8c432fe702201129d853083a314fabf14f6a9e60
Signed-off-by: Matan Barak <matanb@mellanox.com>