IB/uverbs: Use macros to make names consistent in ioctl() uAPI
The ioctl() uAPI works with object-method hierarchy. The method part
also states which handler should be executed when this method is called
from user-space. Therefore, we need to tie method, method's id, method's
handler and the object owning this method together.
Previously, this was done through explicit developer chosen namings.
This makes grepping the code harder. Changing the method's name,
method's handler and object's name to be automatically generated based
on the ids.

issue: 1087238
Change-Id: I971da1149a672342a29c21e1a0588c232346e01c
Signed-off-by: Matan Barak <matanb@mellanox.com>